+++ src/webviews/mongoClusters/collectionView/components/toolbar/ToolbarViewNavigation.tsx
@@ -12,10 +12,9 @@ import { ToolbarDividerTransparent } from './ToolbarDividerTransparent';
 
 export const ToolbarViewNavigation = (): JSX.Element => {
     /**
-     * Use the `useTrpcClient` hook to get the tRPC client and an event target
-     * for handling notifications from the extension.
+     * Use the `useTrpcClient` hook to get the tRPC client
      */
-    const { trpcClient /** , vscodeEventTarget */ } = useTrpcClient();
+    const { trpcClient } = useTrpcClient();
 
     const [currentContext, setCurrentContext] = useContext(CollectionViewContext);
 
@@ -44,7 +43,7 @@ export const ToolbarViewNavigation = (): JSX.Element => {
                 },
             })
             .catch((error) => {
-                console.debug('Failed to report pagination event:', error);
+                console.debug('Failed to report an event:', error);
             });
     }
 
@@ -73,7 +72,7 @@ export const ToolbarViewNavigation = (): JSX.Element => {
                 },
             })
             .catch((error) => {
-                console.debug('Failed to report pagination event:', error);
+                console.debug('Failed to report an event:', error);
             });
     }
 
@@ -97,7 +96,7 @@ export const ToolbarViewNavigation = (): JSX.Element => {
                 },
             })
             .catch((error) => {
-                console.debug('Failed to report pagination event:', error);
+                console.debug('Failed to report an event:', error);
             });
     }
 
@@ -125,7 +124,7 @@ export const ToolbarViewNavigation = (): JSX.Element => {
                 },
             })
             .catch((error) => {
-                console.debug('Failed to report pagination event:', error);
+                console.debug('Failed to report an event:', error);
             });
     }