+++ src/webviews/mongoClusters/documentView/documentView.tsx
@@ -38,10 +38,9 @@ export const DocumentView = (): JSX.Element => {
     const configuration = useConfiguration<DocumentsViewWebviewConfigurationType>();
 
     /**
-     * Use the `useTrpcClient` hook to get the tRPC client and an event target
-     * for handling notifications from the extension.
+     * Use the `useTrpcClient` hook to get the tRPC client
      */
-    const { trpcClient /*, vscodeEventTarget*/ } = useTrpcClient();
+    const { trpcClient } = useTrpcClient();
 
     const [editorContent] = configuration.mode === 'add' ? useState('{  }') : useState('{ "loading...": true }');
     const [isLoading, setIsLoading] = useState(configuration.mode !== 'add');
@@ -53,33 +52,28 @@ export const DocumentView = (): JSX.Element => {
             const documentId: string = configuration.documentId;
 
             setIsLoading(true);
-            void trpcClient.mongoClusters.documentView.getDocumentById.query(documentId).then((response) => {
-                setContent(response);
-            });
 
-            setIsLoading(false);
+            void trpcClient.mongoClusters.documentView.getDocumentById
+                .query(documentId)
+                .then((response) => {
+                    setContent(response);
+                })
+                .catch((error) => {
+                    void trpcClient.common.displayErrorMessage.mutate({
+                        message: 'Error while loading the document',
+                        modal: false,
+                        cause: error instanceof Error ? error.message : String(error),
+                    });
+                })
+                .finally(() => {
+                    setIsLoading(false);
+                });
         }
     }, []);
 
     const editorRef = useRef<monacoEditor.editor.IStandaloneCodeEditor | null>(null);
     const getCurrentContent = () => editorRef.current?.getValue() || '';
     const setContent = (newValue: string) => editorRef.current?.setValue(newValue);
-
-    // useEffect(() => { // example of handling notifications from the extension
-    //     const handleNotification = (event: Event) => {
-    //         const customEvent = event as CustomEvent<VsCodeLinkNotification>;
-    //         const notification = customEvent.detail;
-
-    //         // Handle the notification data, just playing with it
-    //         console.log('Handling notification:', notification);
-    //     };
-
-    //     vscodeEventTarget.addEventListener('VsCodeLinkNotification', handleNotification);
-
-    //     return () => {
-    //         vscodeEventTarget.removeEventListener('VsCodeLinkNotification', handleNotification);
-    //     };
-    // }, [vscodeEventTarget]);
 
     const handleMonacoEditorMount = (
         editor: monacoEditor.editor.IStandaloneCodeEditor,
@@ -180,6 +174,13 @@ export const DocumentView = (): JSX.Element => {
                 documentLength = response.length ?? 0;
                 setContent(response);
             })
+            .catch((error) => {
+                void trpcClient.common.displayErrorMessage.mutate({
+                    message: 'Error while refreshing the document',
+                    modal: false,
+                    cause: error instanceof Error ? error.message : String(error),
+                });
+            })
             .finally(() => {
                 setIsLoading(false);
             });
@@ -195,7 +196,7 @@ export const DocumentView = (): JSX.Element => {
                 },
             })
             .catch((error) => {
-                console.debug('Failed to report event:', error);
+                console.debug('Failed to report an event:', error);
             });
     }
 
@@ -221,7 +222,11 @@ export const DocumentView = (): JSX.Element => {
                 setIsDirty(false);
             })
             .catch((error) => {
-                console.debug('Error saving document:', error);
+                void trpcClient.common.displayErrorMessage.mutate({
+                    message: 'Error saving the document',
+                    modal: true, // we want to show the error in a modal dialog as it's an important one, failed to save the document
+                    cause: error instanceof Error ? error.message : String(error),
+                });
             })
             .finally(() => {
                 setIsLoading(false);
@@ -238,7 +243,7 @@ export const DocumentView = (): JSX.Element => {
                 },
             })
             .catch((error) => {
-                console.debug('Failed to report event:', error);
+                console.debug('Failed to report an event:', error);
             });
     }